feat(ehr): moving files around
Ref: #1040
